arbitrary should maintain the isallmaybe invariant
